# nsqd: per-topic message IDs