@mreiferson 2016-04-17
> One approach is we could narrow the definition of a message ID, to make it more of a ACK id, where it's only purpose is an identifier to FIN/ACK/TOUCH a message, and it's not an ID for the message itself. (i'm not sure i like this, just putting it out there)

Practically speaking that's all it really is and I regret treating them any differently.  Agreed that we need to account for the edge cases you mentioned, but they're pretty easy once you've decided on the above.

By logging I assume you mean in the `nsqd` logs?

Let me get some benchmarks here so we have some numbers to make decisions.