@mreiferson 2016-04-17
Benchmark added and `benchcmp` output below.

NOTE: I think at `16+` my laptop just doesn't have anything left to give

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkProtocolV2PubMultiTopic1-4      1325          1231          -7.09%
BenchmarkProtocolV2PubMultiTopic2-4      890           826           -7.19%
BenchmarkProtocolV2PubMultiTopic4-4      927           633           -31.72%
BenchmarkProtocolV2PubMultiTopic8-4      880           638           -27.50%
BenchmarkProtocolV2PubMultiTopic16-4     745           730           -2.01%
BenchmarkProtocolV2PubMultiTopic32-4     750           685           -8.67%

benchmark                                old MB/s     new MB/s     speedup
BenchmarkProtocolV2PubMultiTopic1-4      150.86       162.44       1.08x
BenchmarkProtocolV2PubMultiTopic2-4      224.62       242.12       1.08x
BenchmarkProtocolV2PubMultiTopic4-4      215.71       315.50       1.46x
BenchmarkProtocolV2PubMultiTopic8-4      227.19       313.21       1.38x
BenchmarkProtocolV2PubMultiTopic16-4     268.36       273.73       1.02x
BenchmarkProtocolV2PubMultiTopic32-4     266.48       291.83       1.10x
```