@judwhite 2016-04-18
I share @jehiah's concerns about this approach. It most noticeably affects nsqd instances which persist every message to disk, and long running work queues, but I believe it could impact in-memory queues also.

Scenario 1:
- nsqd receives message, assigns messageID 1
- Consumer picks up messageID 1
- nsqd restarts before Consumer FIN's the message
- Consumer reconnects to nsqd before attempting to FIN
- nsqd receives a new message, assigns messageID 1
- Original Consumer FIN's the original messageID 1, which actually FIN's the new message.

Scenario 2:
- Message goes to disk without being picked up by a Consumer
- nsqd restarts
- On next PUB, will there be two messages with messageID 1 in the diskqueue?

I think there was always an edge-case problem with using wall-clock to generate the ID since clocks change and cannot be relied upon to be monotonic. There's code to guard against this, but it could make publishing to nsqd unavailable until time proceeds (ever have NTP send you back in time 5 minutes?) - this could actually be another `NOK` scenario. It's also not guaranteed time hasn't gone backwards between nsqd restarts, which the current solution wouldn't be able to guard against.

One possible solution could be to persist the sequence with the metadata. This comes with the same issues as Scenario 1 above for messageID's sent out between persists when nsqd restarts without the ability to sync.

Two ideas around using persisted ID's:
1. If it's purely sequential we'd almost be guaranteeing collisions given the possibility of the persisted ID not being current. A wall-clock component would shorten the possibility of collision, but wouldn't completely eliminate it.
2. To get globally unique ID's I think you'd have to buffer writes for a very short duration, few ms at most, and only reply once the latest sequence has been persisted and synced, allowing you to remove the time component (if you trust fsync). If you specify disk crashes/corruption require assigning a new -worker-id you should have globally unique ID's without the above problems. I'd have to benchmark this idea, I don't have an idea of how much it will impact write throughput and probably _It Depends_™.

My suggestion is a hybrid, start with 1 and work towards 2 if it's feasible, but keep the time component. If NTP and fsync both work against you it's just not your day.

Thoughts?