@mreiferson 2016-04-18
> My biggest concern is that because we expose the ID to the client, people might have mistakenly used it as an ID for the message contents, and this would break that semantics. =(

Yea, but that was always a bad idea.

#625 would, in theory, make the IDs actually useful with stronger guarantees.