@mreiferson 2017-01-02
Last commit restores the time/node based ID generator.

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkProtocolV2PubMultiTopic1-4      597           626           +4.86%
BenchmarkProtocolV2PubMultiTopic2-4      388           392           +1.03%
BenchmarkProtocolV2PubMultiTopic4-4      329           332           +0.91%
BenchmarkProtocolV2PubMultiTopic8-4      360           360           +0.00%
BenchmarkProtocolV2PubMultiTopic16-4     384           387           +0.78%
BenchmarkProtocolV2PubMultiTopic32-4     409           418           +2.20%

benchmark                                old MB/s     new MB/s     speedup
BenchmarkProtocolV2PubMultiTopic1-4      334.67       319.20       0.95x
BenchmarkProtocolV2PubMultiTopic2-4      515.39       509.50       0.99x
BenchmarkProtocolV2PubMultiTopic4-4      607.72       601.08       0.99x
BenchmarkProtocolV2PubMultiTopic8-4      554.42       555.32       1.00x
BenchmarkProtocolV2PubMultiTopic16-4     519.98       516.47       0.99x
BenchmarkProtocolV2PubMultiTopic32-4     488.85       477.89       0.98x
```

Only nominally slower, but still much better scaling characteristics vs `master`:

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkProtocolV2PubMultiTopic1-4      685           626           -8.61%
BenchmarkProtocolV2PubMultiTopic2-4      616           392           -36.36%
BenchmarkProtocolV2PubMultiTopic4-4      578           332           -42.56%
BenchmarkProtocolV2PubMultiTopic8-4      551           360           -34.66%
BenchmarkProtocolV2PubMultiTopic16-4     626           387           -38.18%
BenchmarkProtocolV2PubMultiTopic32-4     514           418           -18.68%

benchmark                                old MB/s     new MB/s     speedup
BenchmarkProtocolV2PubMultiTopic1-4      291.94       319.20       1.09x
BenchmarkProtocolV2PubMultiTopic2-4      324.19       509.50       1.57x
BenchmarkProtocolV2PubMultiTopic4-4      345.50       601.08       1.74x
BenchmarkProtocolV2PubMultiTopic8-4      362.92       555.32       1.53x
BenchmarkProtocolV2PubMultiTopic16-4     319.21       516.47       1.62x
BenchmarkProtocolV2PubMultiTopic32-4     388.86       477.89       1.23x
```